Add files via upload
This is a rough prototype of the form for searching all aspects of the "computer".  It loads into a gridview via a dataAdapter (dataset).
I still have to make it editable and include software info.  But it utilizes the dropdown as we discussed. 
I was using a class to store the results originally, but I don't think it's needed.